support the specified certificate of domain

+++ scripts/oneprovider_init.sh
@@ -1,6 +1,8 @@
 #!/bin/bash -x
 
 apt-get update
+apt-get -y install unzip
+apt-get -y install python
 
 oneprovider_version=`sed -n -e '/^ONEPROVIDER_VERSION/p' /tmp/user-inject.data | awk -F"=" '{print $2}'`
 onezone_domain=`sed -n -e '/^ONEZONE_DOMAIN/p' /tmp/user-inject.data | awk -F"=" '{print $2}'`
@@ -23,25 +25,43 @@ if [[ $oneprovider_version =~ ^18.02.0- ]];then
         ping -c 1 -q $oneprovider_domain
         if [ $? -eq 0 ]
         then
-            #install certbot & create certs
-            apt-get -y install software-properties-common
+            if [ -e /root/obs/aksk.txt ]
+            then
+                python /root/obs/config.py
+                cd /root/
+                unzip onedata.zip
 
-            echo "\n" | add-apt-repository ppa:certbot/certbot
-            apt-get update
-            apt-get -y install certbot
+                mkdir -p /opt/oneprovider/certs
+                cd /opt/oneprovider/certs
+                ln -s /root/onedata/oneprovider/chain.pem chain.pem
+                ln -s /root/onedata/oneprovider/cert.pem cert.pem
+                ln -s /root/onedata/oneprovider/privkey.pem key.pem
 
-            echo 'A' | certbot certonly --standalone -d $oneprovider_domain --register-unsafely-without-email
+                cd /home/ubuntu/onedata/scenarios/3_0_oneprovider_onezone_multihost
+                sed -i 's/#- "${OP_PRIV_KEY_PATH}/- "\/opt\/oneprovider\/certs\/key.pem/' docker-compose-oneprovider.yml
+                sed -i 's/#- "${OP_CERT_PATH}/- "\/opt\/oneprovider\/certs\/cert.pem/' docker-compose-oneprovider.yml
+                sed -i 's/#- "${OZ_CHAIN_PATH}/- "\/opt\/oneprovider\/certs\/chain.pem/g' docker-compose-oneprovider.yml
+            else
+                #install certbot & create certs
+                apt-get -y install software-properties-common
 
-            mkdir -p /opt/oneprovider/certs
-            cd /opt/oneprovider/certs
-            ln -s /etc/letsencrypt/live/$oneprovider_domain/chain.pem chain.pem
-            ln -s /etc/letsencrypt/live/$oneprovider_domain/cert.pem cert.pem
-            ln -s /etc/letsencrypt/live/$oneprovider_domain/privkey.pem key.pem
+                echo "\n" | add-apt-repository ppa:certbot/certbot
+                apt-get update
+                apt-get -y install certbot
 
-            cd /home/ubuntu/onedata/scenarios/3_0_oneprovider_onezone_multihost
-            sed -i 's/#- "${OP_PRIV_KEY_PATH}/- "\/opt\/oneprovider\/certs\/key.pem/' docker-compose-oneprovider.yml
-            sed -i 's/#- "${OP_CERT_PATH}/- "\/opt\/oneprovider\/certs\/cert.pem/' docker-compose-oneprovider.yml
-            sed -i 's/#- "${OZ_CHAIN_PATH}/- "\/opt\/oneprovider\/certs\/chain.pem/g' docker-compose-oneprovider.yml
+                echo 'A' | certbot certonly --standalone -d $oneprovider_domain --register-unsafely-without-email
+
+                mkdir -p /opt/oneprovider/certs
+                cd /opt/oneprovider/certs
+                ln -s /etc/letsencrypt/live/$oneprovider_domain/chain.pem chain.pem
+                ln -s /etc/letsencrypt/live/$oneprovider_domain/cert.pem cert.pem
+                ln -s /etc/letsencrypt/live/$oneprovider_domain/privkey.pem key.pem
+
+                cd /home/ubuntu/onedata/scenarios/3_0_oneprovider_onezone_multihost
+                sed -i 's/#- "${OP_PRIV_KEY_PATH}/- "\/opt\/oneprovider\/certs\/key.pem/' docker-compose-oneprovider.yml
+                sed -i 's/#- "${OP_CERT_PATH}/- "\/opt\/oneprovider\/certs\/cert.pem/' docker-compose-oneprovider.yml
+                sed -i 's/#- "${OZ_CHAIN_PATH}/- "\/opt\/oneprovider\/certs\/chain.pem/g' docker-compose-oneprovider.yml
+            fi
         fi
     fi
 
@@ -58,25 +78,43 @@ else
         ping -c 1 -q $oneprovider_domain
         if [ $? -eq 0 ]
         then
-            #install certbot & create certs
-            apt-get -y install software-properties-common
+            if [ -e /root/obs/aksk.txt ]
+            then
+                python /root/obs/config.py
+                cd /root/
+                unzip onedata.zip
 
-            echo "\n" | add-apt-repository ppa:certbot/certbot
-            apt-get update
-            apt-get -y install certbot
+                mkdir -p /opt/oneprovider/certs
+                cd /opt/oneprovider/certs
+                ln -s /root/onedata/oneprovider/chain.pem cacert.pem
+                ln -s /root/onedata/oneprovider/cert.pem cert.pem
+                ln -s /root/onedata/oneprovider/privkey.pem key.pem
 
-            echo 'A' | certbot certonly --standalone -d $oneprovider_domain --register-unsafely-without-email
+                cd /home/ubuntu/onedata/scenarios/3_0_oneprovider_onezone_multihost
+                sed -i 's/#- "${OP_PRIV_KEY_PATH}/- "\/opt\/oneprovider\/certs\/key.pem/' docker-compose-oneprovider.yml
+                sed -i 's/#- "${OP_CERT_PATH}/- "\/opt\/oneprovider\/certs\/cert.pem/' docker-compose-oneprovider.yml
+                sed -i 's/#- "${OP_CACERT_PATH}/- "\/opt\/oneprovider\/certs\/cacert.pem/g' docker-compose-oneprovider.yml
+            else
+                #install certbot & create certs
+                apt-get -y install software-properties-common
 
-            mkdir -p /opt/oneprovider/certs
-            cd /opt/oneprovider/certs
-            ln -s /etc/letsencrypt/live/$oneprovider_domain/chain.pem cacert.pem
-            ln -s /etc/letsencrypt/live/$oneprovider_domain/fullchain.pem cert.pem
-            ln -s /etc/letsencrypt/live/$oneprovider_domain/privkey.pem key.pem
+                echo "\n" | add-apt-repository ppa:certbot/certbot
+                apt-get update
+                apt-get -y install certbot
 
-            cd /home/ubuntu/onedata/scenarios/3_0_oneprovider_onezone_multihost
-            sed -i 's/#- "${OP_PRIV_KEY_PATH}/- "\/opt\/oneprovider\/certs\/key.pem/' docker-compose-oneprovider.yml
-            sed -i 's/#- "${OP_CERT_PATH}/- "\/opt\/oneprovider\/certs\/cert.pem/' docker-compose-oneprovider.yml
-            sed -i 's/#- "${OP_CACERT_PATH}/- "\/opt\/oneprovider\/certs\/cacert.pem/g' docker-compose-oneprovider.yml
+                echo 'A' | certbot certonly --standalone -d $oneprovider_domain --register-unsafely-without-email
+
+                mkdir -p /opt/oneprovider/certs
+                cd /opt/oneprovider/certs
+                ln -s /etc/letsencrypt/live/$oneprovider_domain/chain.pem cacert.pem
+                ln -s /etc/letsencrypt/live/$oneprovider_domain/fullchain.pem cert.pem
+                ln -s /etc/letsencrypt/live/$oneprovider_domain/privkey.pem key.pem
+
+                cd /home/ubuntu/onedata/scenarios/3_0_oneprovider_onezone_multihost
+                sed -i 's/#- "${OP_PRIV_KEY_PATH}/- "\/opt\/oneprovider\/certs\/key.pem/' docker-compose-oneprovider.yml
+                sed -i 's/#- "${OP_CERT_PATH}/- "\/opt\/oneprovider\/certs\/cert.pem/' docker-compose-oneprovider.yml
+                sed -i 's/#- "${OP_CACERT_PATH}/- "\/opt\/oneprovider\/certs\/cacert.pem/g' docker-compose-oneprovider.yml
+            fi
         fi
     fi
 fi
