+++ scripts/onezone_init.sh
@@ -1,6 +1,8 @@
 #!/bin/bash -x
 
 apt-get update
+apt-get -y install unzip
+apt-get -y install python
 
 onezone_version=`sed -n -e '/^ONEZONE_VERSION/p' /tmp/user-inject.data | awk -F"=" '{print $2}'`
 onezone_domain=`sed -n -e '/^ONEZONE_DOMAIN/p' /tmp/user-inject.data | awk -F"=" '{print $2}'`
@@ -21,25 +23,43 @@ if [[ $onezone_version =~ ^18.02.0- ]];then
         ping -c 1 -q $onezone_domain
         if [ $? -eq 0 ]
         then
-            #install certbot & create certs
-            apt-get -y install software-properties-common
+            if [ -e /root/obs/aksk.txt ]
+            then
+                python /root/obs/config.py
+                cd /root/
+                unzip onedata.zip
 
-            echo "\n" | add-apt-repository ppa:certbot/certbot
-            apt-get update
-            apt-get -y install certbot
+                mkdir -p /opt/onezone/certs
+                cd /opt/onezone/certs
+                ln -s /root/onedata/onezone/chain.pem chain.pem
+                ln -s /root/onedata/onezone/cert.pem cert.pem
+                ln -s /root/onedata/onezone/privkey.pem key.pem
 
-            echo 'A' | certbot certonly --standalone -d $onezone_domain --register-unsafely-without-email
+                cd /home/ubuntu/onedata/scenarios/3_0_oneprovider_onezone_multihost
+                sed -i 's/#- "${OZ_PRIV_KEY_PATH}/- "\/opt\/onezone\/certs\/key.pem/' docker-compose-onezone.yml
+                sed -i 's/#- "${OZ_CERT_PATH}/- "\/opt\/onezone\/certs\/cert.pem/' docker-compose-onezone.yml
+                sed -i 's/#- "${OZ_CHAIN_PATH}/- "\/opt\/onezone\/certs\/chain.pem/g' docker-compose-onezone.yml
+            else
+                #install certbot & create certs
+                apt-get -y install software-properties-common
 
-            mkdir -p /opt/onezone/certs
-            cd /opt/onezone/certs
-            ln -s /etc/letsencrypt/live/$onezone_domain/chain.pem chain.pem
-            ln -s /etc/letsencrypt/live/$onezone_domain/cert.pem cert.pem
-            ln -s /etc/letsencrypt/live/$onezone_domain/privkey.pem key.pem
+                echo "\n" | add-apt-repository ppa:certbot/certbot
+                apt-get update
+                apt-get -y install certbot
 
-            cd /home/ubuntu/onedata/scenarios/3_0_oneprovider_onezone_multihost
-            sed -i 's/#- "${OZ_PRIV_KEY_PATH}/- "\/opt\/onezone\/certs\/key.pem/' docker-compose-onezone.yml
-            sed -i 's/#- "${OZ_CERT_PATH}/- "\/opt\/onezone\/certs\/cert.pem/' docker-compose-onezone.yml
-            sed -i 's/#- "${OZ_CHAIN_PATH}/- "\/opt\/onezone\/certs\/chain.pem/g' docker-compose-onezone.yml
+                echo 'A' | certbot certonly --standalone -d $onezone_domain --register-unsafely-without-email
+
+                mkdir -p /opt/onezone/certs
+                cd /opt/onezone/certs
+                ln -s /etc/letsencrypt/live/$onezone_domain/chain.pem chain.pem
+                ln -s /etc/letsencrypt/live/$onezone_domain/cert.pem cert.pem
+                ln -s /etc/letsencrypt/live/$onezone_domain/privkey.pem key.pem
+
+                cd /home/ubuntu/onedata/scenarios/3_0_oneprovider_onezone_multihost
+                sed -i 's/#- "${OZ_PRIV_KEY_PATH}/- "\/opt\/onezone\/certs\/key.pem/' docker-compose-onezone.yml
+                sed -i 's/#- "${OZ_CERT_PATH}/- "\/opt\/onezone\/certs\/cert.pem/' docker-compose-onezone.yml
+                sed -i 's/#- "${OZ_CHAIN_PATH}/- "\/opt\/onezone\/certs\/chain.pem/g' docker-compose-onezone.yml
+            fi
         fi
     fi
 
@@ -55,25 +75,43 @@ else
         ping -c 1 -q $onezone_domain
         if [ $? -eq 0 ]
         then
-            #install certbot & create certs
-            apt-get -y install software-properties-common
+            if [ -e /root/obs/aksk.txt ]
+            then
+                python /root/obs/config.py
+                cd /root/
+                unzip onedata.zip
 
-            echo "\n" | add-apt-repository ppa:certbot/certbot
-            apt-get update
-            apt-get -y install certbot
+                mkdir -p /opt/onezone/certs
+                cd /opt/onezone/certs
+                ln -s /root/onedata/onezone/chain.pem cacert.pem
+                ln -s /root/onedata/onezone/cert.pem cert.pem
+                ln -s /root/onedata/onezone/privkey.pem key.pem
 
-            echo 'A' | certbot certonly --standalone -d $onezone_domain --register-unsafely-without-email
+                cd /home/ubuntu/onedata/scenarios/3_0_oneprovider_onezone_multihost
+                sed -i 's/#- "${OZ_PRIV_KEY_PATH}/- "\/opt\/onezone\/certs\/key.pem/' docker-compose-onezone.yml
+                sed -i 's/#- "${OZ_CERT_PATH}/- "\/opt\/onezone\/certs\/cert.pem/' docker-compose-onezone.yml
+                sed -i 's/#- "${OZ_CACERT_PATH}/- "\/opt\/onezone\/certs\/cacert.pem/g' docker-compose-onezone.yml
+            else
+                #install certbot & create certs
+                apt-get -y install software-properties-common
 
-            mkdir -p /opt/onezone/certs
-            cd /opt/onezone/certs
-            ln -s /etc/letsencrypt/live/$onezone_domain/chain.pem cacert.pem
-            ln -s /etc/letsencrypt/live/$onezone_domain/fullchain.pem cert.pem
-            ln -s /etc/letsencrypt/live/$onezone_domain/privkey.pem key.pem
+                echo "\n" | add-apt-repository ppa:certbot/certbot
+                apt-get update
+                apt-get -y install certbot
 
-            cd /home/ubuntu/onedata/scenarios/3_0_oneprovider_onezone_multihost
-            sed -i 's/#- "${OZ_PRIV_KEY_PATH}/- "\/opt\/onezone\/certs\/key.pem/' docker-compose-onezone.yml
-            sed -i 's/#- "${OZ_CERT_PATH}/- "\/opt\/onezone\/certs\/cert.pem/' docker-compose-onezone.yml
-            sed -i 's/#- "${OZ_CACERT_PATH}/- "\/opt\/onezone\/certs\/cacert.pem/g' docker-compose-onezone.yml
+                echo 'A' | certbot certonly --standalone -d $onezone_domain --register-unsafely-without-email
+
+                mkdir -p /opt/onezone/certs
+                cd /opt/onezone/certs
+                ln -s /etc/letsencrypt/live/$onezone_domain/chain.pem cacert.pem
+                ln -s /etc/letsencrypt/live/$onezone_domain/fullchain.pem cert.pem
+                ln -s /etc/letsencrypt/live/$onezone_domain/privkey.pem key.pem
+
+                cd /home/ubuntu/onedata/scenarios/3_0_oneprovider_onezone_multihost
+                sed -i 's/#- "${OZ_PRIV_KEY_PATH}/- "\/opt\/onezone\/certs\/key.pem/' docker-compose-onezone.yml
+                sed -i 's/#- "${OZ_CERT_PATH}/- "\/opt\/onezone\/certs\/cert.pem/' docker-compose-onezone.yml
+                sed -i 's/#- "${OZ_CACERT_PATH}/- "\/opt\/onezone\/certs\/cacert.pem/g' docker-compose-onezone.yml
+            fi
         fi
     fi
 fi
